Repaired parallel trainer
