+++ lib/natural/classifiers/classifier.js
@@ -27,15 +27,6 @@ const events = require('events')
 
 const parallelTrainer = require('./classifier_train_parallel')
 
-let Threads = null
-
-try {
-  Threads = require('webworker-threads')
-} catch (e) {
-  // Since webworker-threads are optional, only thow if the module is found
-  if (e.code !== 'MODULE_NOT_FOUND') throw e
-}
-
 const Classifier = function (classifier, stemmer) {
   this.classifier = classifier
   this.docs = []
@@ -111,25 +102,6 @@ function textToFeatures (observation) {
   return features
 }
 
-function docsToFeatures (docs) {
-  const parsedDocs = []
-
-  for (let i = 0; i < docs.length; i++) {
-    const features = []
-
-    for (const feature in FEATURES) { // eslint-disable-line
-      if (docs[i].observation.indexOf(feature) > -1) { features.push(1) } else { features.push(0) }
-    }
-
-    parsedDocs.push({
-      index: docs[i].index,
-      features: features
-    })
-  }
-
-  return JSON.stringify(parsedDocs)
-}
-
 function train () {
   const totalDocs = this.docs.length
   for (let i = this.lastAdded; i < totalDocs; i++) {
@@ -196,7 +168,7 @@ function setOptions (options) {
 Classifier.prototype.addDocument = addDocument
 Classifier.prototype.removeDocument = removeDocument
 Classifier.prototype.train = train
-if (Threads) {
+if (parallelTrainer.Threads) {
   Classifier.prototype.trainParallel = parallelTrainer.trainParallel
   Classifier.prototype.trainParallelBatches = parallelTrainer.trainParallelBatches
   Classifier.prototype.retrainParallel = parallelTrainer.retrainParallel
