+++ lib/natural/classifiers/classifier_train_parallel.js
@@ -1,8 +1,64 @@
 const os = require('os')
 
+let Threads = null
+
+function docsToFeatures (docs) {
+  const parsedDocs = []
+
+  for (let i = 0; i < docs.length; i++) {
+    const features = []
+
+    for (const feature in FEATURES) { // eslint-disable-line
+      if (docs[i].observation.indexOf(feature) > -1) { features.push(1) } else { features.push(0) }
+    }
+
+    parsedDocs.push({
+      index: docs[i].index,
+      features: features
+    })
+  }
+
+  return JSON.stringify(parsedDocs)
+}
+
+try {
+  Threads = require('webworker-threads')
+} catch (e) {
+  // Since webworker-threads are optional, only thow if the module is found
+  if (e.code !== 'MODULE_NOT_FOUND') throw e
+}
+
 function checkThreadSupport () {
   if (typeof Threads === 'undefined') {
     throw new Error('parallel classification requires the optional dependency webworker-threads')
+  }
+}
+
+// Convert docs to observation objects
+function docsToObs (docs, lastAdded, stemmer, keepStops) {
+  const obsDocs = []
+  for (let i = lastAdded; i < docs.length; i++) {
+    let observation = this.docs[i].text
+    if (typeof observation === 'string') {
+      observation = stemmer.tokenizeAndStem(observation, keepStops)
+    }
+    obsDocs.push({
+      index: i,
+      observation: observation
+    })
+  }
+  return obsDocs
+}
+
+function emitEvents (self, docFeatures, totalDocs) {
+  for (let j = self.lastAdded; j < totalDocs; j++) {
+    self.classifier.addExample(docFeatures[j], self.docs[j].label)
+    self.events.emit('trainedWithDocument', {
+      index: j,
+      total: totalDocs,
+      doc: self.docs[j]
+    })
+    self.lastAdded++
   }
 }
 
@@ -28,16 +84,7 @@ function trainParallel (numThreads, callback) {
   threadPool.all.eval('var FEATURES = ' + JSON.stringify(this.features))
   threadPool.all.eval(docsToFeatures)
 
-  // Convert docs to observation objects
-  const obsDocs = []
-  for (let i = this.lastAdded; i < totalDocs; i++) {
-    let observation = this.docs[i].text
-    if (typeof observation === 'string') { observation = this.stemmer.tokenizeAndStem(observation, this.keepStops) }
-    obsDocs.push({
-      index: i,
-      observation: observation
-    })
-  }
+  const obsDocs = docsToObs(this.docs, this.lastAdded, this.stemmer, this.keepStops)
 
   // Called when a batch completes processing
   const onFeaturesResult = function (docs) {
@@ -62,16 +109,7 @@ function trainParallel (numThreads, callback) {
       return callback(err)
     }
 
-    for (let j = self.lastAdded; j < totalDocs; j++) {
-      self.classifier.addExample(docFeatures[j], self.docs[j].label)
-      self.events.emit('trainedWithDocument', {
-        index: j,
-        total: totalDocs,
-        doc: self.docs[j]
-      })
-      self.lastAdded++
-    }
-
+    emitEvents(self, docFeatures, totalDocs)
     self.events.emit('doneTraining', true)
     self.classifier.train()
 
@@ -181,16 +219,7 @@ function trainParallelBatches (options) {
     threadPool.destroy(true)
     abort = true
 
-    for (let j = self.lastAdded; j < totalDocs; j++) {
-      self.classifier.addExample(docFeatures[j], self.docs[j].label)
-      self.events.emit('trainedWithDocument', {
-        index: j,
-        total: totalDocs,
-        doc: self.docs[j]
-      })
-      self.lastAdded++
-    }
-
+    emitEvents(self, docFeatures, totalDocs)
     self.events.emit('doneTraining', true)
     self.classifier.train()
   }
@@ -243,6 +272,7 @@ function retrainParallel (numThreads, callback) {
 }
 
 module.exports = {
+  Threads,
   trainParallel,
   trainParallelBatches,
   retrainParallel
